// backup-src/types.ts
export interface IUser {
  _id: string;
  username: string; // unique identifyer
  pi_uid: string; // hashed user password
  brand?: string; // User Legal Name (e.g. Tony Adeola Ezenwa) (optional)
  image?: string; // URL to user profile pics (optional)
  email?: string; // for notification
  phone?: string; // user phone number (optional)
  provider?: string; // URLs of reviewer upload (optional)
};
// ...
export enum UserTypeEnum {
  Individual = 'individual',
  Company = 'company',
  Agent = 'agent',
}

export interface UserSettingsType {
  _id?: string;
  username: string; // 
  user_type: UserTypeEnum;
  brand?: string; // User Legal Name (e.g. Tony Adeola Ezenwa)
  image?: string; // URL to user profile pics (optional)
  email?: string; // for notification    
  phone?: string; // URLs of reviewer upload (optional)
  whatsapp?: string; // User whatsapp number
  social_handles?: {platform: string; handle: string}[]; // e.g [{platform: 'facebook', handle: 'fb.com/username'}]
};

export enum CurrencyEnum {
  naira = "NGN",
  dollars="USD",
  pounds = "GBP",
  euros = "EUR",
}

export enum CategoryEnum {
  house = "house",
  shortlet="shortlet",
  hotel = "hotel",
  office = "office",
  land = "land",
  shop = "shop",
  others = "others"
}

export enum ListForEnum {
  rent = "rent",
  sale = "sale"
}

export enum RenewalEnum {
  monthly = "monthly",
  yearly = "yearly",
  daily = "daily",
  weekely = "weekly"
}

export enum NegotiableEnum {
  Negotiable = "negotiable",
  NonNegotiable = "Non-negotiable",
}

export enum PropertyStatusEnum {
  available = "available",
  sold = "sold",
  rented = "rented",
  unavailable = "unavailable",
  expired = "expired",
}

export interface PropertyType {
  id: string;
  banner: string; // URL of the property image or image with index = 0
  title: string; // Title of the property (e.g. 3 bedroom flat, self contain, )
  slug: string;
  address: string; // Location of the property
  price: number; // Price per month
  currency: CurrencyEnum; // Currency type (e.g. NGN, USD, GBP)
  listed_for: string; // (e.g. "sell"/ "rent")
  category: CategoryEnum; // The class of property (e.g. house, land, shop, office, hotel )
  period?: RenewalEnum; // if is for rent, payment period (e.g monthly, yearly, daily)
  negotiable: boolean; // (true/false)
  property_terms?: string // agent's term cond
  images: string[]; //Other property images for gallery
  user: IUser; //foreign key representing agent that list the property
  latitude: number;
  longitude: number;
  features?: {
    name: string;
    quantity: number;
  }[];
  env_facilities?: string[];
  rating?: number;
  status: PropertyStatusEnum; // (available, sold, unavailable, rented)
  createdAt?: Date;
  updatedAt?: Date;
};